refactor website
