+++ site/component/utils.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import hljs from 'highlight.js';
-import antd, { Menu } from '../../';
+import antd from '../../';
 
 function isHeading(type) {
   return /h[1-6]/i.test(type);
@@ -62,33 +62,4 @@ export function objectToComponent(object, index) {
     object.type, { key: index },
     children && children.map(objectToComponent) // `hr` has no children
   );
-}
-
-export function flattenMenu(menu) {
-  if (menu.type === Menu.Item) {
-    return menu;
-  }
-
-  if (Array.isArray(menu)) {
-    return menu.reduce((acc, item) => {
-      return acc.concat(flattenMenu(item));
-    }, []);
-  }
-
-  return flattenMenu(menu.props.children);
-}
-
-export function getActiveMenuItem(props, index) {
-  const routes = props.routes;
-  return routes[routes.length - 1].path || index;
-}
-
-export function getFooterNav(menuItems, activeMenuItem) {
-  const menuItemsList = flattenMenu(menuItems);
-  const activeMenuItemIndex = menuItemsList.findIndex((menuItem) => {
-    return menuItem.key === activeMenuItem;
-  });
-  const prev = menuItemsList[activeMenuItemIndex - 1];
-  const next = menuItemsList[activeMenuItemIndex + 1];
-  return { prev, next };
 }
